Начало библиотеки для свечей
# Conflicts:
#	app/templates/crypto_currency.html
#	instance/database.db
#	migrations/__pycache__/env.cpython-313.pyc
#	migrations/versions/5e963451ad67_.py
#	migrations/versions/__pycache__/e1dad00c06f4_initial_migration.cpython-313.pyc
#	migrations/versions/c62daf0f458b_initial.py
#	migrations/versions/e1dad00c06f4_initial_migration.py

+++ app/models.py
@@ -4,7 +4,7 @@ from werkzeug.security import generate_password_hash, check_password_hash
 from wtforms import StringField, PasswordField, SubmitField, validators
 import ccxt
 import time
-from sqlalchemy import Table, Column, Integer, Float, String, LargeBinary
+from sqlalchemy import Table, Column, Integer, Float, String, LargeBinary, BigInteger
 import re
 from flask_wtf.file import FileField, FileAllowed
 from flask_wtf import FlaskForm
@@ -42,7 +42,7 @@ class OHLCV(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     symbol = db.Column(db.String)
     timeframe = db.Column(db.String)
-    timestamp = db.Column(db.Integer)
+    timestamp = db.Column(db.BigInteger)
     datetime = db.Column(db.String)
     open = db.Column(db.Float)
     high = db.Column(db.Float)
